Updated loading bar to take in estimated completion time; minor refactoring
